# lib/api-zod/src/generated/types/transcribeQuestionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TranscribeQuestionInput {
  /**
   * @minLength 1
   * @maxLength 6291456
   */
  imageDataUrl: string;
  /** @maxLength 100 */
  subject?: string;
}