# lib/api-zod/src/generated/types/transcribeQuestionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TranscribeQuestionResult {
  transcription: string;
  readable: boolean;
}